Move into a metadata class.
