+++ airbyte-commons-worker/src/main/java/io/airbyte/workers/sync/LauncherWorker.java
@@ -9,7 +9,7 @@ import static io.airbyte.metrics.lib.ApmTraceConstants.Tags.JOB_ID_KEY;
 import static io.airbyte.metrics.lib.ApmTraceConstants.Tags.JOB_ROOT_KEY;
 import static io.airbyte.metrics.lib.ApmTraceConstants.Tags.PROCESS_EXIT_VALUE_KEY;
 import static io.airbyte.metrics.lib.ApmTraceConstants.WORKER_OPERATION_NAME;
-import static io.airbyte.workers.process.KubeProcessFactory.CONNECTION_ID_LABEL_KEY;
+import static io.airbyte.workers.process.Metadata.CONNECTION_ID_LABEL_KEY;
 
 import com.google.common.base.Stopwatch;
 import datadog.trace.api.Trace;
